task 0 fix error

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,17 +8,17 @@
 
 size_t print_list(const list_t *h)
 {
-    long unsigned int n = 0;
+	unsigned int n = 0;
 
-    if (h == NULL)
-    return (-1);
+	if (h == NULL)
+	return (-1);
 
-    while (h != NULL)
-    {
-        n++;
-        printf("[%d] %s\n", h->len, h->str);
-        h = h->next;
-    }
+	while (h != NULL)
+	{
+	n++;
+	printf("[%d] %s\n", h->len, h->str);
+	h = h->next;
+	}
 
-    return (n);
+	return (n);
 }
